Removed unsafe AsynchronousSocketGroup from UDP API
